refactor(AAS-14): add more structure and rename to export
